requirements.txt
aiohttp                                  3.9.5
aiosignal                                1.3.1
altair                                   5.3.0
anyio                                    3.7.1
asgiref                                  3.8.1
asttokens                                2.4.1
attrs                                    23.2.0
backoff                                  2.2.1
bcrypt                                   4.1.3
blinker                                  1.8.2
build                                    1.2.1
cachetools                               5.3.3
certifi                                  2024.7.4
charset-normalizer                       3.3.2
chroma-hnswlib                           0.7.3
chromadb                                 0.5.3
click                                    8.1.7
collection                               0.1.6
colorama                                 0.4.6
coloredlogs                              15.0.1
comm                                     0.2.2
contourpy                                1.2.1
cycler                                   0.12.1
dataclasses-json                         0.6.7
debugpy                                  1.8.2
decorator                                5.1.1
Deprecated                               1.2.14
distro                                   1.9.0
embeddings                               0.0.8
executing                                2.0.1
fastapi                                  0.95.2
filelock                                 3.15.4
flatbuffers                              24.3.25
fonttools                                4.53.1
frozenlist                               1.4.1
fsspec                                   2024.6.1
gitdb                                    4.0.11
GitPython                                3.1.43
google-auth                              2.32.0
googleapis-common-protos                 1.63.2
greenlet                                 3.0.3
grpcio                                   1.64.1
h11                                      0.14.0
haversine                                2.8.1
httpcore                                 1.0.5
httptools                                0.6.1
httpx                                    0.27.0
huggingface-hub                          0.23.4
humanfriendly                            10.0
idna                                     3.7
importlib_metadata                       7.1.0
importlib_resources                      6.4.0
ipykernel                                6.29.5
ipython                                  8.26.0
jedi                                     0.19.1
Jinja2                                   3.1.4
joblib                                   1.4.2
jsonpatch                                1.33
jsonpointer                              3.0.0
jsonschema                               4.23.0
jsonschema-specifications                2023.12.1
jupyter_client                           8.6.2
jupyter_core                             5.7.2
kiwisolver                               1.4.5
kubernetes                               30.1.0
langchain                                0.2.7
langchain-community                      0.2.7
langchain-core                           0.2.12
langchain-text-splitters                 0.2.2
langsmith                                0.1.84
markdown-it-py                           3.0.0
MarkupSafe                               2.1.5
marshmallow                              3.21.3
matplotlib                               3.9.1
matplotlib-inline                        0.1.7
mdurl                                    0.1.2
mmh3                                     4.1.0
monotonic                                1.6
mpmath                                   1.3.0
multidict                                6.0.5
mypy-extensions                          1.0.0
nest-asyncio                             1.6.0
numpy                                    1.26.4
oauthlib                                 3.2.2
onnxruntime                              1.18.1
openai                                   1.35.14
opentelemetry-api                        1.25.0
opentelemetry-exporter-otlp-proto-common 1.25.0
opentelemetry-exporter-otlp-proto-grpc   1.25.0
opentelemetry-instrumentation            0.46b0
opentelemetry-instrumentation-asgi       0.46b0
opentelemetry-instrumentation-fastapi    0.46b0
opentelemetry-proto                      1.25.0
opentelemetry-sdk                        1.25.0
opentelemetry-semantic-conventions       0.46b0
opentelemetry-util-http                  0.46b0
orjson                                   3.10.6
overrides                                7.7.0
packaging                                24.1
pandas                                   2.2.2
parso                                    0.8.4
pillow                                   10.4.0
pip                                      24.1.2
platformdirs                             4.2.2
posthog                                  3.5.0
prompt_toolkit                           3.0.47
protobuf                                 4.25.3
psutil                                   6.0.0
pure-eval                                0.2.2
pyarrow                                  16.1.0
pyasn1                                   0.6.0
pyasn1_modules                           0.4.0
pydantic                                 1.10.9
pydeck                                   0.9.1
Pygments                                 2.18.0
pyparsing                                3.1.2
PyPika                                   0.48.9
pyproject_hooks                          1.1.0
pyreadline3                              3.4.1
python-dateutil                          2.9.0.post0
python-dotenv                            1.0.1
pytz                                     2024.1
pywin32                                  306
PyYAML                                   6.0.1
pyzmq                                    26.0.3
referencing                              0.35.1
requests                                 2.32.3
requests-oauthlib                        2.0.0
rich                                     13.7.1
rpds-py                                  0.19.0
rsa                                      4.9
scikit-learn                             1.5.1
scipy                                    1.14.0
setuptools                               65.5.0
shellingham                              1.5.4
six                                      1.16.0
smmap                                    5.0.1
sniffio                                  1.3.1
SQLAlchemy                               2.0.31
stack-data                               0.6.3
starlette                                0.27.0
streamlit                                1.36.0
sympy                                    1.13.0
tenacity                                 8.5.0
threadpoolctl                            3.5.0
tokenizers                               0.19.1
toml                                     0.10.2
toolz                                    0.12.1
tornado                                  6.4.1
tqdm                                     4.66.4
traitlets                                5.14.3
typer                                    0.12.3
typing_extensions                        4.12.2
typing-inspect                           0.9.0
tzdata                                   2024.1
urllib3                                  2.2.2
uvicorn                                  0.22.0
watchdog                                 4.0.1
watchfiles                               0.22.0
wcwidth                                  0.2.13
websocket-client                         1.8.0
websockets                               12.0
wheel                                    0.43.0
wrapt                                    1.16.0
yarl                                     1.9.4
zipp                                     3.19.2